Removed dialog module - moved DateTimeUtils to app/src/kotlin/base
