+++ app/src/main/java/bose/ankush/weatherify/presentation/home/HomeViewModel.kt
@@ -6,20 +6,18 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import bose.ankush.weatherify.R
-import bose.ankush.weatherify.base.AssetLoader
+import bose.ankush.weatherify.base.DateTimeUtils.getDayNameFromEpoch
 import bose.ankush.weatherify.common.Extension.getForecastListForNext4Days
 import bose.ankush.weatherify.common.ResultData
 import bose.ankush.weatherify.common.UiText
 import bose.ankush.weatherify.data.remote.dto.ForecastDto
 import bose.ankush.weatherify.domain.model.AirQuality
 import bose.ankush.weatherify.domain.model.AvgForecast
-import bose.ankush.weatherify.domain.model.Country
 import bose.ankush.weatherify.domain.model.Weather
 import bose.ankush.weatherify.domain.use_case.get_air_quality.GetAirQuality
 import bose.ankush.weatherify.domain.use_case.get_weather_forecasts.GetForecasts
 import bose.ankush.weatherify.domain.use_case.get_weather_reports.GetTodaysWeatherReport
 import bose.ankush.weatherify.presentation.UIState
-import com.bosankus.utilities.DateTimeUtils.getDayNameFromEpoch
 import com.google.android.gms.location.FusedLocationProviderClient
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
