+++ settings.gradle.kts
@@ -18,4 +18,4 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "Weatherify"
-include (":app", ":utilities", ":payment", ":language")
+include (":app", ":payment", ":language")
